% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_and_extract_zip}
\alias{download_and_extract_zip}
\title{Title Downloads and extracts external data to be used in the simulation}
\usage{
download_and_extract_zip(url, dest_folder = ".", zip_name = NULL)
}
\arguments{
\item{url}{the URL from where the external data can be downloaded. The external data has be wrapped in a zip file}

\item{dest_folder}{The folder where the data should be stored}

\item{zip_name}{Only needs to be changed if the zip file is differently named as the last part of the URL}
}
\value{
nothing, stores the data on the local system
}
\description{
Title Downloads and extracts external data to be used in the simulation
}
\examples{
\dontrun{
download_and_extract_zip(url = "www.nextcloud.de/mysuperfile")
}
}
