% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdesign_utils.R
\name{extract_b_values}
\alias{extract_b_values}
\title{Title Extracts beta values from an spdesign object}
\usage{
extract_b_values(input_list)
}
\arguments{
\item{input_list}{the list where the parameters are stored. Usually this is \code{design$utility}}
}
\value{
A named list with parameter values which can be used in \code{sim_all}
}
\description{
Title Extracts beta values from an spdesign object
}
\examples{
\dontrun{
extract_b_values(design$utility)
}
}
