% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateDCE-package.R
\docType{package}
\name{simulateDCE-package}
\alias{simulateDCE}
\alias{simulateDCE-package}
\title{simulateDCE: Simulate Data for Discrete Choice Experiments}
\description{
Supports simulating choice experiment data for given designs. It helps to quickly test different designs against each other and compare the performance of new models. The goal of 'simulateDCE' is to make it easy to simulate choice experiment datasets using designs from 'NGENE', 'idefix' or 'spdesign'. You have to store the design file(s) in a sub-directory and need to specify certain parameters and the utility functions for the data generating process. For more details on choice experiments see Mariel et al. (2021) \doi{10.1007/978-3-030-62669-3}.
}
\author{
\strong{Maintainer}: Julian Sagebiel \email{julian.sagebiel@idiv.de} (\href{https://orcid.org/0000-0002-0253-6875}{ORCID})

}
\keyword{internal}
