% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateResults.R
\name{aggregateResults}
\alias{aggregateResults}
\title{Aggregate Simulation Results}
\usage{
aggregateResults(all_designs)
}
\arguments{
\item{all_designs}{A list of simulation results from sim_choice. Can contain different designs but need to have the common structure returned by simchoice}
}
\value{
A list with aggregated results including summary, coefficients, graphs, and power.
}
\description{
Processes the simulation results to extract summaries, coefficients, and graphs.
}
