% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdesign.R
\name{readdesign}
\alias{readdesign}
\title{Creates a dataframe with the design.}
\usage{
readdesign(design = designfile, designtype = destype)
}
\arguments{
\item{design}{The path to a design file}

\item{designtype}{Is it a design created with ngene or with spdesign. Ngene desings should be stored as the standard .ngd output. spdesign should be the spdesign object design$design}
}
\value{
a dataframe
}
\description{
Creates a dataframe with the design.
}
\examples{
\dontrun{readdesign("Projects/test/Designs/firstdesign.ngd", "ngene")}

}
