% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_all.R
\name{sim_all}
\alias{sim_all}
\title{Title Is a wrapper for sim_choice executing the simulation over all designs stored in a specific folder}
\usage{
sim_all(nosim = 2, resps, destype = "ngene", designpath)
}
\arguments{
\item{nosim}{Number of runs or simulations. For testing use 2 but once you go serious, use at least 200, for better results use 2000.}

\item{resps}{Number of respondents you want  to simulate}

\item{destype}{Is it a design created with ngene or with spdesign. Ngene desings should be stored as the standard .ngd output. spdesign should be the spdesign object design$design}

\item{designpath}{The path to the folder where the designs are stored. For example "c:/myfancydec/Designs"}
}
\value{
a list, with all information on the simulation. This list an be easily processed by the user and in the rmarkdown template.
}
\description{
Title Is a wrapper for sim_choice executing the simulation over all designs stored in a specific folder
}
\examples{

 designpath<- system.file("extdata","Rbook" ,package = "simulateDCE")
 resps =240  # number of respondents
 nosim=2 # number of simulations to run (about 500 is minimum)

}
