% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_choice.R
\name{sim_choice}
\alias{sim_choice}
\title{Title}
\usage{
sim_choice(
  designfile,
  no_sim = 10,
  respondents = 330,
  mnl_U,
  utils = u[[1]],
  destype
)
}
\arguments{
\item{destype}{}
}
\description{
Title
}
