% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_choice.R
\name{sim_choice}
\alias{sim_choice}
\title{Simulate and estimate choices}
\usage{
sim_choice(
  designfile,
  no_sim = 10,
  respondents = 330,
  ut,
  designtype = NULL,
  destype = NULL,
  bcoefficients,
  decisiongroups = c(0, 1),
  manipulations = list(),
  estimate,
  chunks = 1,
  utility_transform_type = "simple"
)
}
\arguments{
\item{designfile}{path to a file containing a design.}

\item{no_sim}{Number of runs i.e. how often do you want the simulation to be repeated}

\item{respondents}{Number of respondents. How many respondents do you want to simulate in each run.}

\item{ut}{The first element of the utility function list}

\item{designtype}{Is it a design created with ngene, spdesign or idefix. use 'ngene', 'spdesign' or 'idefix. Ngene designs should be stored as the standard .ngd output. spdesign should be the spdesign object stored as an RDS file. Idefix objects should also be stored as an RDS file. If designtype is not specified, I try to guess what it is. This is especially helpful if you want to carry out a simulation for both spdesign designs and ngene designs at the same time.}

\item{destype}{Deprecated. Use designtype instead.}

\item{bcoefficients}{List of initial coefficients for the utility function. List content/length can vary based on application, but should all begin with b and be the same as those entered in the utility functions}

\item{decisiongroups}{A vector showing how decision groups are numerically distributed}

\item{manipulations}{A variable to alter terms of the utility functions examples may be applying a factor or applying changes to terms selectively for different groups}

\item{estimate}{If TRUE models will be estimated. If false only a dataset will be simulated. Default is true}

\item{chunks}{The number of chunks determines how often results should be stored on disk as a safety measure to not loose simulations if models have already been estimated. For example, if no_sim is 100 and chunks = 2, the data will be saved on disk after 50 and after 100 runs.}

\item{utility_transform_type}{How the utility function you entered is transformed to the utility function required for mixl. You can use the classic way (simple) where parameters have to start with "b" and variables with "alt" or the more flexible (but potentially error prone) way (exact) where parameters and variables are matched exactly what how the are called in the dataset and in the bcoeff list. Default is "simple". In the long run, simple will be deleted, as exact should be downwards compatible.}
}
\value{
a list with all information on the run
}
\description{
Simulate and estimate choices
}
\examples{
\dontrun{  simchoice(designfile="somefile", no_sim=10, respondents=330,
 mnl_U,ut=u[[1]] ,designtype="ngene")}

}
