% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_choice.R
\name{sim_choice}
\alias{sim_choice}
\title{Simulate and estimate choices}
\usage{
sim_choice(
  designfile,
  no_sim = 10,
  respondents = 330,
  ut,
  destype = destype,
  bcoefficients,
  decisiongroups = c(0, 1),
  manipulations = list(),
  estimate
)
}
\arguments{
\item{designfile}{path to a file containing a design.}

\item{no_sim}{Number of runs i.e. how often do you want the simulation to be repeated}

\item{respondents}{Number of respondents. How many respondents do you want to simulate in each run.}

\item{ut}{The first element of the utility function list}

\item{bcoefficients}{List of initial coefficients for the utility function. List content/length can vary based on application, but should all begin with b and be the same as those entered in the utility functions}

\item{decisiongroups}{A vector showing how decision groups are numerically distributed}

\item{manipulations}{A variable to alter terms of the utility functions examples may be applying a factor or applying changes to terms selectively for different groups}

\item{estimate}{If TRUE models will be estimated. If false only a dataset will be simulated. Default is true}
}
\value{
a list with all information on the run
}
\description{
Simulate and estimate choices
}
\examples{
\dontrun{  simchoice(designfile="somefile", no_sim=10, respondents=330,
 mnl_U,ut=u[[1]] ,destype="ngene")}

}
