% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_choices.R
\name{simulate_choices}
\alias{simulate_choices}
\title{Simulate choices based on a dataframe with a design}
\usage{
simulate_choices(data = datadet, utility = utils, setspp)
}
\arguments{
\item{data}{a dataframe that includes a design repeated for the number of observations}

\item{utility}{a list with the utility functions, one utility function for each alternatives}

\item{setspp}{an integeger, the number of choice sets per person}
}
\value{
a dataframe that includes simulated choices and a design
}
\description{
Simulate choices based on a dataframe with a design
}
\examples{
\dontrun{simulate_choices(datadet, utils,setspp)}
}
