% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_choices.R
\name{simulate_choices}
\alias{simulate_choices}
\title{Simulate choices based on a dataframe with a design}
\usage{
simulate_choices(
  data,
  utility,
  setspp,
  destype,
  bcoefficients,
  decisiongroups = c(0, 1),
  manipulations = list(),
  estimate
)
}
\arguments{
\item{data}{a dataframe that includes a design repeated for the number of observations}

\item{utility}{a list with the utility functions, one utility function for each alternatives}

\item{setspp}{an integer, the number of choice sets per person}

\item{bcoefficients}{List of initial coefficients for the utility function. List content/length can vary based on application, but should all begin with b and be the same as those entered in the utility functions}

\item{decisiongroups}{A vector showing how decision groups are numerically distributed}

\item{manipulations}{A variable to alter terms of the utility functions examples may be applying a factor or applying changes to terms selectively for different groups}
}
\value{
a dataframe that includes simulated choices and a design
}
\description{
Simulate choices based on a dataframe with a design
}
\examples{
\dontrun{simulate_choices(datadet, ut,setspp)}
}
