% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_choices.R
\name{simulate_choices}
\alias{simulate_choices}
\title{Simulate choices based on a dataframe with a design}
\usage{
simulate_choices(data, utility, setspp, destype)
}
\arguments{
\item{data}{a dataframe that includes a design repeated for the number of observations}

\item{utility}{a list with the utility functions, one utility function for each alternatives}

\item{setspp}{an integer, the number of choice sets per person}

\item{destype}{Is it a design created with ngene or with spdesign. Ngene desings should be stored as the standard .ngd output. spdesign should be the spdesign object design$design}
}
\value{
a dataframe that includes simulated choices and a design
}
\description{
Simulate choices based on a dataframe with a design
}
\examples{
\dontrun{simulate_choices(datadet, ut,setspp)}
}
